Moved most of .run-httpd.py code to new httpd/__init__.py
